phibIA-app/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography==41.0.7
numpy
librosa
scikit-learn
soundfile
flask-bcrypt==1.0.1 
Flask-JWT-Extended
flask-cors
